+++ src/main/java/com/abclinic/server/repository/AlbumRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface AlbumRepository extends JpaRepository<ImageAlbum, Integer> {
-    Optional<ImageAlbum> findById(int id);
+public interface AlbumRepository extends JpaRepository<ImageAlbum, Long> {
+    Optional<ImageAlbum> findById(long id);
     Optional<List<ImageAlbum>> findByPatient(Patient patient);
 }
