+++ src/main/java/com/abclinic/server/repository/DietitianRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface DietitianRepository extends JpaRepository<Dietitian, Integer> {
-    Optional<Dietitian> findById(int id);
+public interface DietitianRepository extends JpaRepository<Dietitian, Long> {
+    Optional<Dietitian> findById(long id);
 }
