+++ src/main/java/com/abclinic/server/repository/ImageRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface ImageRepository extends JpaRepository<Image, Integer> {
-    Optional<Image> findById(int id);
+public interface ImageRepository extends JpaRepository<Image, Long> {
+    Optional<Image> findById(long id);
 }
