+++ src/main/java/com/abclinic/server/repository/MedicalRecordRepository.java
@@ -10,8 +10,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface MedicalRecordRepository extends JpaRepository<MedicalRecord, Integer> {
-    Optional<MedicalRecord> findById(int id);
+public interface MedicalRecordRepository extends JpaRepository<MedicalRecord, Long> {
+    Optional<MedicalRecord> findById(long id);
     Optional<List<MedicalRecord>> findByPatient(Patient patient);
     Optional<List<MedicalRecord>> findByDisease(Disease disease);
     Optional<List<MedicalRecord>> findByPractitioner(Practitioner practitioner);
