+++ src/main/java/com/abclinic/server/repository/QuestionRepository.java
@@ -3,6 +3,6 @@ package com.abclinic.server.repository;
 import com.abclinic.server.model.entity.Question;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuestionRepository extends JpaRepository<Question, Integer> {
+public interface QuestionRepository extends JpaRepository<Question, Long> {
 
 }
