+++ src/main/java/com/abclinic/server/repository/ReplyRepository.java
@@ -3,5 +3,5 @@ package com.abclinic.server.repository;
 import com.abclinic.server.model.entity.Reply;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReplyRepository extends JpaRepository<Reply, Integer> {
+public interface ReplyRepository extends JpaRepository<Reply, Long> {
 }
