+++ src/main/java/com/abclinic/server/repository/SpecialistRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface SpecialistRepository extends JpaRepository<Specialist, Integer> {
-    Optional<Specialist> findById(int id);
+public interface SpecialistRepository extends JpaRepository<Specialist, Long> {
+    Optional<Specialist> findById(long id);
     Optional<List<Specialist>> findBySpecialty(Specialty specialty);
 }
