+++ src/main/java/com/abclinic/server/repository/SpecialtyRepository.java
@@ -8,6 +8,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * @package com.abclinic.server.repository
  * @created 12/27/2019 4:11 PM
  */
-public interface SpecialtyRepository extends JpaRepository<Specialty, Integer> {
+public interface SpecialtyRepository extends JpaRepository<Specialty, Long> {
 
 }
